fix: create seperate component for playback
