+++ app/containers/ITunes/index.js
@@ -4,14 +4,14 @@
  *
  */
 
-import React, { useEffect, useState, useRef } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { injectIntl } from 'react-intl';
 import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
 import { injectSaga } from 'redux-injectors';
-import { Card, Input, Row, Skeleton, Progress } from 'antd';
+import { Card, Input, Row, Skeleton } from 'antd';
 import { selectITunesData, selectITunesError, selectITunesName, selectITunesLoading } from './selectors';
 import saga from './saga';
 import styled from 'styled-components';
@@ -21,7 +21,7 @@ import If from '@components/If';
 import For from '@components/For';
 import T from '@components/T';
 import { AlbumCard } from './components/AlbumCard';
-import { PlayCircleFilled, PauseCircleFilled } from '@ant-design/icons';
+import PlayBack from './components/PlayBack/index';
 
 const { Search } = Input;
 const CustomCard = styled(Card)`
@@ -40,34 +40,6 @@ const Container = styled.div`
     width: 100%;
     margin: 0 auto;
     padding: ${(props) => props.padding}px;
-  }
-`;
-const PlayDiv = styled.div`
-  && {
-    background: black;
-    color: white;
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    justify-content: space-between;
-    position: fixed;
-    bottom: 0;
-    width: 690px;
-    border-radius: 10px 10px 0px 0px;
-    box-shadow: 0px 0px 10px 8px #888888;
-  }
-`;
-const PlayAlbumInfo = styled.div`
-  && {
-    display: flex;
-    align-items: center;
-    margin-left: 15px;
-  }
-`;
-const PlayAlbumControl = styled.div`
-  && {
-    width: 35%;
-    margin-right: 35px;
   }
 `;
 export function ITunes({
@@ -146,55 +118,6 @@ export function ITunes({
       )
     );
   };
-  const playSong = () => {
-    let audioRef = useRef('');
-    if (currentSong.songName) {
-      // console.log(audioRef.current.currentTime());
-      return (
-        <PlayDiv>
-          <PlayAlbumInfo>
-            <img src={currentSong.img} />
-            <T text={currentSong.songName} style={{ marginLeft: '10px' }} />
-          </PlayAlbumInfo>
-          <If
-            condition={currentSong.play}
-            otherwise={
-              <PlayCircleFilled
-                style={{ fontSize: '25px' }}
-                onClick={() => {
-                  audioRef.current.play();
-                  setCurrentSong({ ...currentSong, play: true });
-                }}
-              />
-            }
-          >
-            <PauseCircleFilled
-              style={{ fontSize: '25px' }}
-              onClick={() => {
-                audioRef.current.pause();
-                setCurrentSong({ ...currentSong, play: false });
-              }}
-            />
-          </If>
-          <PlayAlbumControl>
-            <Progress percent={currentSong.progress} status="active" size="small" showInfo={false} />
-          </PlayAlbumControl>
-          <audio
-            autoPlay
-            src={currentSong.previewUrl}
-            ref={audioRef}
-            onEnded={() => setCurrentSong({ ...currentSong, play: false, progress: 0 })}
-            onTimeUpdate={() =>
-              setCurrentSong({
-                ...currentSong,
-                progress: (audioRef.current.currentTime / audioRef.current.duration) * 100
-              })
-            }
-          />
-        </PlayDiv>
-      );
-    }
-  };
   return (
     <Container maxwidth="700" padding="10">
       <CustomCard title={intl.formatMessage({ id: 'album_search' })}>
@@ -209,7 +132,9 @@ export function ITunes({
       </CustomCard>
       {renderList(iTunesData)}
       {renderError()}
-      {playSong()}
+      <If condition={currentSong.songName}>
+        <PlayBack currentSong={currentSong} setCurrentSong={setCurrentSong} />
+      </If>
     </Container>
   );
 }
